Update update_daemon_json.php
Added missing null check before reloading Dockerd

+++ source/kata.runtime.ui/kata.runtime/scripts/update_daemon_json.php
@@ -60,8 +60,11 @@
   file_put_contents($file_path, $newJsonString);
 
   //Trigger Dockerd to reload the daemon file
-  echo "Reloading Dockerd...\n";
   $pidstr = shell_exec("pidof dockerd");
-  $pidint = intval($pidstr);
-  posix_kill($pidint, SIGHUP);
+  if( !($pidstr == NULL) )
+  {
+    echo "Reloading Dockerd...\n";  
+    $pidint = intval($pidstr);
+    posix_kill($pidint, SIGHUP);
+  }
 ?>
